# improve code krnln_rnd.cpp