- updated krnln_rnd.cpp
- removed unuseful comment

---

**Performance Information** _(100'000'000 times loop)_

> AMD Ryzen 7 1700 Eight-Core Processor 3.00Ghz
> OS Windows 10 64-bit (17763), 32 GB RAM 

`取随机数 (0, 2147483647)`

| Type  | 1  | 2  | 3  |
| -- | -- | -- | -- |
| old  | 1984 ms  | 2016 ms  | 2047 ms |
| new  | 1781 ms  | 1765 ms  | 1782 ms  |

`取随机数 (-1, -1)`

| Type  | 1  | 2  | 3  |
| -- | -- | -- | -- |
| old  | 375 ms  | 391 ms  | 391 ms |
| new  | 359 ms  | 391 ms  | **407 ms**  |

`取随机数 (-1, 233333)`

| Type  | 1  | 2  | 3  |
| -- | -- | -- | -- |
| old  | 1953 ms  | 1953 ms  | 1906 ms |
| new  | 1813 ms  | 1797 ms  | 1797 ms  |